booking_service/requirements.txt
django==5.0.14
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
kafka-python==2.0.2
psycopg2-binary==2.9.9
requests==2.31.0
redis==5.0.1
django-redis==5.4.0
